requirements/requirements-dev.txt
pytest
ruff
numpy
pillow
matplotlib
